requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
